+++ activity/src/main/java/com/kairos/service/counter/CounterHelperService.java
@@ -88,7 +88,7 @@ public class CounterHelperService {
         return new Object[]{staffIds,filterDates,unitIds,employmentTypeIds,daysOfWeeks,shiftActivityStatus,plannedTimeIds};
     }
 
-    public Set<DayOfWeek> getDayOfWeek(List<Long> dayTypeIds,Map<Long, DayTypeDTO> daysTypeIdAndDayTypeMap)
+    public Set<DayOfWeek> getDayOfWeek(List<BigInteger> dayTypeIds,Map<BigInteger, DayTypeDTO> daysTypeIdAndDayTypeMap)
     {
         Set<DayOfWeek> daysOfWeek = new HashSet<>();
